User decorator for discount curve and set up new SpreadFittingMethod fitting method
Yes, it makes it better to use the decorator. Now the change is fully
backwards compatible, and besides it looks better as only
SpreadFittingMethod knows about an underlying discount curve, instead of
that being contaminated everywhere. I added some getter methods to the
base class FittingMethod, so as not to have to pass the parameters
twice, once to the calculation method and once to the spread helper.